requirements.txt
streamlit==1.28.0
#openbb==4.1.4
numpy==1.26.4
matplotlib
datetime
pandas
yfinance
backtesting
datetime
ta-lib-bin
seaborn
openpyxl
plotly
st-pages
scipy
